Pixel alpha is an unsigned int in the 0-256 range.

+++ src/a2x_pack_blit.v.h
@@ -30,5 +30,5 @@ extern void a_blit__init(void);
 extern void a_blit__setBlend(APixelBlend Blend);
 extern void a_blit__setClip(bool DoClip);
 
-extern void a_blit__setAlpha(uint8_t Alpha);
+extern void a_blit__setAlpha(unsigned int Alpha);
 extern void a_blit__setRGB(uint8_t Red, uint8_t Green, uint8_t Blue);
